+++ src/main/kotlin/jasition/matching/domain/book/TradingStatus.kt
@@ -1,5 +1,6 @@
 package jasition.matching.domain.book
 
+import jasition.matching.domain.order.command.CancelOrderCommand
 import jasition.matching.domain.order.command.PlaceOrderCommand
 import jasition.matching.domain.quote.command.CancelMassQuoteCommand
 import jasition.matching.domain.quote.command.PlaceMassQuoteCommand
@@ -16,31 +17,37 @@ data class TradingStatuses(
 enum class TradingStatus {
     OPEN_FOR_TRADING {
         override fun allows(command: PlaceMassQuoteCommand): Boolean = true
+        override fun allows(command: CancelOrderCommand): Boolean = true
         override fun allows(command: PlaceOrderCommand): Boolean = true
         override fun allows(command: CancelMassQuoteCommand): Boolean = true
     },
     HALTED {
         override fun allows(command: PlaceMassQuoteCommand): Boolean = false
+        override fun allows(command: CancelOrderCommand): Boolean = true
         override fun allows(command: PlaceOrderCommand): Boolean = false
         override fun allows(command: CancelMassQuoteCommand): Boolean = true
     },
     NOT_AVAILABLE_FOR_TRADING {
         override fun allows(command: PlaceMassQuoteCommand): Boolean = false
+        override fun allows(command: CancelOrderCommand): Boolean = true
         override fun allows(command: PlaceOrderCommand): Boolean = false
         override fun allows(command: CancelMassQuoteCommand): Boolean = true
     },
     PRE_OPEN {
         override fun allows(command: PlaceMassQuoteCommand): Boolean = true
+        override fun allows(command: CancelOrderCommand): Boolean = true
         override fun allows(command: PlaceOrderCommand): Boolean = false
         override fun allows(command: CancelMassQuoteCommand): Boolean = true
     },
     SYSTEM_MAINTENANCE {
         override fun allows(command: PlaceOrderCommand): Boolean = false
+        override fun allows(command: CancelOrderCommand): Boolean = false
         override fun allows(command: PlaceMassQuoteCommand): Boolean = false
         override fun allows(command: CancelMassQuoteCommand): Boolean = false
     };
 
     abstract fun allows(command: PlaceOrderCommand): Boolean
+    abstract fun allows(command: CancelOrderCommand): Boolean
     abstract fun allows(command: PlaceMassQuoteCommand): Boolean
     abstract fun allows(command: CancelMassQuoteCommand): Boolean
 }