+++ src/main/kotlin/jasition/matching/domain/book/entry/EntryStatus.kt
@@ -12,6 +12,9 @@ enum class EntryStatus {
     },
     CANCELLED {
         override fun isFinal(): Boolean = true
+    },
+    REJECTED {
+        override fun isFinal(): Boolean = true
     };
 
     abstract fun isFinal(): Boolean;