+++ src/test/kotlin/jasition/matching/domain/order/command/PlaceOrderCommandTest.kt
@@ -107,7 +107,7 @@ internal class PlaceOrderCommandTest : StringSpec({
                 ),
                 events = list(
                     orderPlacedEvent,
-                    expectedOrderCancelledByExchangeEvent(
+                    expectedOrderCancelledEvent(
                         bookId = bookId,
                         eventId = EventId(2),
                         entry = orderPlacedEvent.toBookEntry()